Split inventorization views
